Remove compile warnings from example code. refs #27

+++ examples/c/ex_call_center.c
@@ -139,8 +139,8 @@ int main()
 
 	count = 0;
 	while (call.cust_id == cust.id) {
-		printf("Got call record on date %llu: type %s: %s\n",
-		    call.call_date, call.call_type, call.notes);
+		printf("Got call record on date %lu: type %s: %s\n",
+		    (unsigned long)call.call_date, call.call_type, call.notes);
 		if (++count == 3)
 			break;
